Rework how sources are stored
`sources.json` used to be very machine-oriented, which made it annoying
to edit. This commit changes that by reversing the order from
`{domain: name}` to `{name: domain}` and allowing `domain` to be
a list of domains. The translation to `{domain: name}` still occurs, but inside
`load-sources`.

A new script `tidy_sources` replaces `regenerate_sources.py`
and its wrapper. `tidy_sources` automatically diagnoses and fixes problems
with `sources.json`.